server: telnet: fix comparison between signed and unsigned warning
Exposed with clang 4.0.0.

Since tc->line_cursor is a signed int (-1 is (ab)used to suppress prompt
during log output) and pos is unsigned as required by the semantics of
this function an explicit cast is needed to suppress the warning when
comparing the two.

+++ src/server/telnet_server.c
@@ -310,7 +310,7 @@ static void telnet_move_cursor(struct connection *connection, size_t pos)
 
 	tc = connection->priv;
 
-	if (pos < tc->line_cursor) {
+	if ((int)pos < tc->line_cursor) {
 		tmp = tc->line_cursor - pos;
 
 		for (size_t i = 0; i < tmp; i += 16)
